refactoring state-chip
