+++ web/libs/app-common/src/components/state-chips/state-history-popover-content.tsx
@@ -1,0 +1,94 @@
+/**
+ * StateHistoryPopoverContent - Popover content for displaying state history
+ */
+
+import { Badge } from "@humansignal/ui";
+import { IconSync, IconError, IconHistoryRewind } from "@humansignal/icons";
+import { useStateHistory, type StateHistoryItem } from "../../hooks/useStateHistory";
+import { getStateColorClass, formatStateName, formatTimestamp, formatUserName } from "./utils";
+
+export interface StateHistoryPopoverContentProps {
+  entityType: "task" | "annotation" | "project";
+  entityId: number;
+  isOpen: boolean;
+}
+
+export function StateHistoryPopoverContent({ entityType, entityId, isOpen }: StateHistoryPopoverContentProps) {
+  const { data, isLoading, isError, error, refetch } = useStateHistory({
+    entityType,
+    entityId,
+    enabled: isOpen,
+  });
+
+  const history = (data?.results || []) as StateHistoryItem[];
+
+  return (
+    <div
+      className="flex flex-col w-[320px] max-h-[400px] bg-white dark:bg-gray-900 rounded-lg shadow-lg"
+      onClick={(e) => e.stopPropagation()}
+    >
+      {/* Header */}
+      <div className="px-4 py-3 border-b border-gray-200 dark:border-gray-700">
+        <div className="flex items-center gap-2">
+          <IconHistoryRewind className="w-4 h-4 text-gray-600 dark:text-gray-400" />
+          <span className="text-sm font-medium text-gray-900 dark:text-gray-100">State History</span>
+        </div>
+      </div>
+
+      {/* Content */}
+      <div className="flex-1 overflow-y-auto p-4">
+        {isLoading && (
+          <div className="flex flex-col items-center justify-center py-8 gap-3">
+            <IconSync className="w-8 h-8 text-blue-500 animate-spin" />
+            <span className="text-sm text-gray-500">Loading...</span>
+          </div>
+        )}
+
+        {isError && (
+          <div className="flex flex-col items-center justify-center py-8 gap-3">
+            <IconError className="w-8 h-8 text-red-500" />
+            <span className="text-sm text-gray-900 dark:text-gray-100">Failed to load history</span>
+            <span className="text-xs text-gray-500 text-center">
+              {error instanceof Error ? error.message : "Unknown error"}
+            </span>
+            <button
+              onClick={(e) => {
+                e.stopPropagation();
+                refetch();
+              }}
+              className="mt-2 px-3 py-1.5 text-xs font-medium text-white bg-blue-600 hover:bg-blue-700 rounded transition-colors"
+              type="button"
+            >
+              Retry
+            </button>
+          </div>
+        )}
+
+        {!isLoading && !isError && history.length === 0 && (
+          <div className="flex flex-col items-center justify-center py-8 gap-3">
+            <IconHistoryRewind className="w-8 h-8 text-gray-400" />
+            <span className="text-sm text-gray-500">No history available</span>
+          </div>
+        )}
+
+        {!isLoading && !isError && history.length > 0 && (
+          <div className="space-y-3">
+            {history.map((item: StateHistoryItem, index: number) => (
+              <div key={index} className="pb-3 border-b border-gray-100 dark:border-gray-800 last:border-0 last:pb-0">
+                <div className="flex items-center justify-between mb-2">
+                  <Badge className={getStateColorClass(item.state)}>{formatStateName(item.state)}</Badge>
+                  <span className="text-xs text-gray-500">{formatTimestamp(item.created_at)}</span>
+                </div>
+                <div className="text-xs text-gray-600 dark:text-gray-400">
+                  <div>By: {formatUserName(item.triggered_by)}</div>
+                  {item.transition_name && <div className="mt-1 text-gray-500">{item.transition_name}</div>}
+                </div>
+              </div>
+            ))}
+          </div>
+        )}
+      </div>
+    </div>
+  );
+}
+
