+++ web/libs/app-common/src/index.ts
@@ -4,3 +4,6 @@ export { pages };
 
 // Hooks
 export { useStateHistory, type StateHistoryItem, type StateHistoryResponse } from "./hooks/useStateHistory";
+
+// Components
+export * from "./components/state-chips";
