+++ web/libs/ui/src/lib/state-chip/index.ts
@@ -1,0 +1,2 @@
+export { StateChip, type StateChipProps } from "./state-chip";
+
